feat(console): time to style this
