--- console/client/src/components/Modal.tsx
@@ -1,71 +1,0 @@
-import React from 'react'
-import { Dialog, Transition } from '@headlessui/react'
-
-
-export const Modal: React.FC<{
-  trigger?: React.ElementType
-  title?: string
-  children: React.ReactNode
-  defaultOpen?: boolean
-}> = ({ trigger: Trigger, title, children, defaultOpen = false }) => {
-  const [ isOpen, setIsOpen ] = React.useState(defaultOpen)
-
-  function closeModal() {
-    setIsOpen(false)
-  }
-
-  function openModal() {
-    setIsOpen(true)
-  }
-
-  return (
-    <>
-      {Trigger && <Trigger onClick={openModal} />}
-      <Transition appear
-        show={isOpen}
-        as={React.Fragment}
-      >
-        <Dialog as='div'
-          className='relative z-10'
-          onClose={closeModal}
-        >
-          <Transition.Child
-            as={React.Fragment}
-            enter='ease-out duration-300'
-            enterFrom='opacity-0'
-            enterTo='opacity-100'
-            leave='ease-in duration-200'
-            leaveFrom='opacity-100'
-            leaveTo='opacity-0'
-          >
-            <div className='fixed inset-0 bg-black bg-opacity-25' />
-          </Transition.Child>
-
-          <div className='fixed inset-0 overflow-y-auto'>
-            <div className='flex min-h-full items-center justify-center p-4 text-center'>
-              <Transition.Child
-                as={React.Fragment}
-                enter='ease-out duration-300'
-                enterFrom='opacity-0 scale-95'
-                enterTo='opacity-100 scale-100'
-                leave='ease-in duration-200'
-                leaveFrom='opacity-100 scale-100'
-                leaveTo='opacity-0 scale-95'
-              >
-                <Dialog.Panel className={`w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all`}>
-                  <Dialog.Title
-                    as='h3'
-                    className='text-lg font-medium leading-6 text-gray-900'
-                  >
-                    {title}
-                  </Dialog.Title>
-                  {children}
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
-    </>
-  )
-}
