+++ console/client/src/features/modules/Module.module.css.d.ts
@@ -3,4 +3,7 @@ export const grid: string;
 export const filter: string;
 export const timeline: string;
 export const misc: string;
+export const modules: string;
+export const radio: string;
+export const radioComment: string;
